Moving constants, making winprob and amount global constants
