+++ packages/core/src/constants.ts
@@ -1,12 +1,6 @@
-import BN from 'bn.js'
-
 const pkg = require('../package.json')
 
-export const PRICE_PER_PACKET = '10000000000000000' // 0.01 HOPR
-// Must be a natural number, will be rounded to a natural number otherwise
-export const INVERSE_TICKET_WIN_PROB = '1' // 100%
 export const PACKET_SIZE = 500
-
 export const FULL_VERSION = pkg.version
 
 const packageVersion = pkg.version.split('.')
@@ -34,13 +28,6 @@ export const MAX_HOPS = 2
 export const PATH_RANDOMNESS = 0.1
 export const MAX_PATH_ITERATIONS = 100
 export const NETWORK_QUALITY_THRESHOLD = 0.5
-export const MINIMUM_REASONABLE_CHANNEL_STAKE = new BN(PRICE_PER_PACKET).muln(100)
 export const MAX_NEW_CHANNELS_PER_TICK = 5
-export const MAX_AUTO_CHANNELS = 5
-export const MIN_NATIVE_BALANCE = new BN('1000000000000000') // 0.001 ETH
-
-// enough to fund 10 channels
-export const SUGGESTED_BALANCE = MINIMUM_REASONABLE_CHANNEL_STAKE.muln(MAX_AUTO_CHANNELS * 2)
-export const SUGGESTED_NATIVE_BALANCE = MIN_NATIVE_BALANCE.muln(250) // 0.025 ETH
 
 export const CHECK_TIMEOUT = 10000
