+++ packages/hoprd/src/admin.ts
@@ -1,4 +1,4 @@
-import Hopr, { SUGGESTED_NATIVE_BALANCE } from '@hoprnet/hopr-core'
+import Hopr from '@hoprnet/hopr-core'
 import BN from 'bn.js'
 import http from 'http'
 import fs from 'fs'
@@ -11,7 +11,7 @@ import type { Server } from 'http'
 import stripAnsi from 'strip-ansi'
 import { LogStream } from './logs'
 import { NODE_ENV } from './env'
-import { Balance, NativeBalance } from '@hoprnet/hopr-utils'
+import { Balance, NativeBalance, SUGGESTED_NATIVE_BALANCE } from '@hoprnet/hopr-utils'
 import { Commands } from './commands'
 
 let debugLog = debug('hoprd:admin')
